% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CITestStat}
\alias{CITestStat}
\title{Estimate the CI test statistics}
\usage{
CITestStat(X, Y, Z, h)
}
\arguments{
\item{X}{A matrix of covariates}

\item{Y}{A vector of response}

\item{Z}{A vector of conditional variable}

\item{h}{kernel bandwidth}
}
\description{
Estimate the CI test statistics
}
