% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimateRho}
\alias{EstimateRho}
\title{Estimate rho}
\usage{
EstimateRho(U, V, W)
}
\arguments{
\item{U}{A vector of U}

\item{V}{A vector of V}

\item{W}{A vector of W}
}
\description{
Estimate rho
}
